requirements.txt
psycopg2
telegram.ext
configparser
